+++ app/src/main/java/cmpt276/jade/carbontracker/Route_List_Activity.java
@@ -113,8 +113,8 @@ public class Route_List_Activity extends AppCompatActivity {
                     String RouteName = data.getStringExtra("pass back the route name");
                     String StrHighWay = data.getStringExtra("pass back the highway");
                     String StrCity = data.getStringExtra("pass back the city");
-                    int highway = Integer.parseInt(StrHighWay);
-                    int city = Integer.parseInt(StrCity);
+                    double highway = Double.parseDouble(StrHighWay);
+                    double city =  Double.parseDouble(StrCity);
 
                     Route addedRoute = new Route(RouteName, highway, city);
 
@@ -139,8 +139,8 @@ public class Route_List_Activity extends AppCompatActivity {
                         String StrHighWay = data.getStringExtra("pass back the highway");
                         String StrCity = data.getStringExtra("pass back the city");
 
-                        int highway = Integer.parseInt(StrHighWay);
-                        int city = Integer.parseInt(StrCity);
+                        double highway =  Double.parseDouble(StrHighWay);
+                        double city =  Double.parseDouble(StrCity);
 
                         Route Clicked = new Route(RouteName, highway, city);
                         routes.changeRoute(Clicked, index);
